Node 5
Node 5 is a disorganized mess, its colors represent a bright world while simultaneously appealing to a very young demographic of children. The spawns should be representative of imagination, fantasy, future or other fictacious environments and not a tropic. Take node 2 and its theme of video games. It contains the most details of any spawn while also not mentioning any partisan topics.